Refactor: Logo.png & Login - SignUp Button
